+++ src/avz_smart.cpp
@@ -2,7 +2,7 @@
  * @coding: utf-8
  * @Author: vector-wlc
  * @Date: 2020-02-06 10:22:46
- * @Description: CLASS ACobManager
+ * @Description:
  */
 #include "avz_smart.h"
 #include "avz_asm.h"